+++ CondCore/ESSources/src/PedestalRetriever.cc
@@ -8,7 +8,7 @@
 //
 // Author:      Chris Jones
 // Created:     Mon Apr 18 16:43:35 EDT 2005
-// $Id: PedestalRetriever.cc,v 1.1 2005/07/07 13:19:37 xiezhen Exp $
+// $Id: PedestalRetriever.cc,v 1.1 2005/07/22 14:24:58 xiezhen Exp $
 //
 
 // system include files
@@ -135,7 +135,7 @@ cond::PedestalRetriever::setIntervalFor( const EventSetupRecordKey&,
 				   edm::ValidityInterval& oValidity)
 {
   std::cout<<"PedestalRetriever::setIntervalFor "<< iTime.value()<<std::endl;
-  typedef std::map<unsigned long, std::string> IOVMap;
+  typedef std::map<int, std::string> IOVMap;
   typedef IOVMap::const_iterator iterator;
   try{
     unsigned long abtime=iTime.value()-edm::Timestamp::beginOfTime().value();
